+++ js/main.js
@@ -1,7 +1,5 @@
-// Как правильнее подключить pristine? Здесь, или в index.html?
-// import '../vendor/pristine/pristine.min.js';
 import { fillPictures } from './minis.js';
-import './modal-photo-edit.js';
-import './validation.js';
+import './modal/photo-edit.js';
+import './validation/validation.js';
 
 fillPictures();
